requirements.txt
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
streamlit
mcp
python-dotenv
